--- packages/apollo-engine-reporting/.npmignore
@@ -1,6 +1,0 @@
-*
-!src/**/*
-!dist/**/*
-dist/**/*.test.*
-!package.json
-!README.md
