--- packages/apollo-engine-reporting/README.md
@@ -1,6 +1,0 @@
-# apollo-engine-reporting
-
-[![npm version](https://badge.fury.io/js/apollo-engine-reporting.svg)](https://badge.fury.io/js/apollo-engine-reporting)
-[![Build Status](https://circleci.com/gh/apollographql/apollo-server/tree/main.svg?style=svg)](https://circleci.com/gh/apollographql/apollo-server)
-
-This package is a pure JS implementation of the Apollo Graph Manager (previously, Apollo Engine) reporting feature. It enables aggregated reporting in stateful environments and sending single reports for stateless environments.
