--- packages/apollo-engine-reporting/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "apollo-engine-reporting",
-  "version": "2.3.0",
-  "description": "Send reports about your GraphQL services to Apollo Graph Manager (previously known as Apollo Engine)",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "license": "MIT",
-  "repository": "https://github.com/apollographql/apollo-engine-reporting",
-  "author": "Apollo <opensource@apollographql.com>",
-  "engines": {
-    "node": ">=6.0"
-  },
-  "dependencies": {
-    "apollo-reporting-protobuf": "file:../apollo-reporting-protobuf",
-    "apollo-graphql": "^0.6.0",
-    "apollo-server-caching": "file:../apollo-server-caching",
-    "apollo-server-env": "file:../apollo-server-env",
-    "apollo-server-errors": "file:../apollo-server-errors",
-    "apollo-server-plugin-base": "file:../apollo-server-plugin-base",
-    "apollo-server-types": "file:../apollo-server-types",
-    "async-retry": "^1.2.1",
-    "uuid": "^8.0.0"
-  },
-  "peerDependencies": {
-    "graphql": "^0.12.0 || ^0.13.0 || ^14.0.0 || ^15.0.0"
-  }
-}
