--- packages/apollo-engine-reporting/src/index.ts
@@ -1,1 +1,0 @@
-export { EngineReportingOptions, EngineReportingAgent } from './agent';
