--- packages/apollo-engine-reporting/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../../tsconfig.base",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./dist"
-  },
-  "include": ["src/**/*"],
-  "exclude": ["**/__tests__", "**/__mocks__"],
-  "references": [
-    { "path": "../apollo-server-errors" },
-    { "path": "../apollo-server-types" },
-    { "path": "../apollo-server-plugin-base" },
-  ]
-}
