+++ packages/apollo-server-core/src/plugin/index.ts
@@ -2,7 +2,8 @@ export {
   ApolloServerPluginSchemaReporting,
   ApolloServerPluginSchemaReportingOptions,
 } from './schemaReporting';
+export * from './usageReporting';
 export {
   ApolloServerPluginInlineTrace,
-  ApolloServerPluginInlineTraceOptions,
+  ApolloServerPluginInlineTraceDisabled,
 } from './inlineTrace';
