+++ packages/apollo-server-core/src/plugin/usageReporting/__tests__/signatureCache.test.ts
@@ -1,0 +1,13 @@
+import { signatureCacheKey } from '../signatureCache';
+
+describe('signature cache key', () => {
+  it('generates without the operationName', () => {
+    expect(signatureCacheKey('abc123', '')).toEqual('abc123');
+  });
+
+  it('generates without the operationName', () => {
+    expect(signatureCacheKey('abc123', 'myOperation')).toEqual(
+      'abc123:myOperation',
+    );
+  });
+});
