+++ packages/apollo-server-core/src/plugin/usageReporting/index.ts
@@ -1,0 +1,6 @@
+export {
+  ApolloServerPluginUsageReporting,
+  ApolloServerPluginUsageReportingDisabled,
+} from './plugin';
+export * from './options';
+export { EngineReportingOptions } from './legacyOptions';
