+++ packages/apollo-server-core/tsconfig.json
@@ -9,7 +9,6 @@
   "references": [
     { "path": "../apollo-cache-control" },
     { "path": "../apollo-datasource" },
-    { "path": "../apollo-engine-reporting" },
     { "path": "../apollo-server-caching" },
     { "path": "../apollo-server-errors" },
     { "path": "../apollo-server-plugin-base" },
