+++ packages/apollo-server-plugin-operation-registry/src/__tests__/ApolloServerPluginOperationRegistry.test.ts
@@ -1,5 +1,8 @@
 import plugin, { Options } from '../ApolloServerPluginOperationRegistry';
-import { ApolloServerBase } from 'apollo-server-core';
+import {
+  ApolloServerBase,
+  ApolloServerPluginUsageReportingDisabled,
+} from 'apollo-server-core';
 import {
   /**
    * We alias these to different names entirely since the user-facing values
@@ -21,7 +24,6 @@ import {
 } from './helpers.test-helpers';
 import { Headers } from 'apollo-server-env';
 import { GraphQLRequest } from 'apollo-server-plugin-base';
-import { EngineReportingOptions } from 'apollo-engine-reporting';
 
 // While not ideal, today, Apollo Server has a very real expectation of an HTTP
 // request context.  That will change in the future.  While we can sometimes
@@ -52,10 +54,9 @@ describe('Operation registry plugin', () => {
     const graphId = 'test-service';
     const apiKey = `service:${graphId}:not-an-api-key`;
     const hashedApiKey = hashApiKey(apiKey);
-    const engineOptions: EngineReportingOptions<any> = {
-      apiKey,
-      reportTiming: false,
-      graphVariant: 'current',
+    const apollo: ApolloConfigInput = {
+      key: apiKey,
+      variant: 'current',
     };
     const typeDefs = gql`
       type Query {
@@ -94,8 +95,9 @@ describe('Operation registry plugin', () => {
         const server = new ApolloServerMock({
           typeDefs,
           mockEntireSchema: true,
-          engine: engineOptions,
+          apollo,
           plugins: [
+            ApolloServerPluginUsageReportingDisabled(),
             plugin({
               onUnregisteredOperation,
             })(),
@@ -144,8 +146,9 @@ describe('Operation registry plugin', () => {
         const server = new ApolloServerMock({
           typeDefs,
           mockEntireSchema: true,
-          engine: engineOptions,
+          apollo,
           plugins: [
+            ApolloServerPluginUsageReportingDisabled(),
             plugin({
               onUnregisteredOperation,
             })(),
@@ -180,8 +183,9 @@ describe('Operation registry plugin', () => {
         const server = new ApolloServerMock({
           typeDefs,
           mockEntireSchema: true,
-          engine: engineOptions,
+          apollo,
           plugins: [
+            ApolloServerPluginUsageReportingDisabled(),
             plugin({
               forbidUnregisteredOperations,
               onForbiddenOperation,
@@ -233,8 +237,9 @@ describe('Operation registry plugin', () => {
         const server = new ApolloServerMock({
           typeDefs,
           mockEntireSchema: true,
-          engine: engineOptions,
+          apollo,
           plugins: [
+            ApolloServerPluginUsageReportingDisabled(),
             plugin({
               forbidUnregisteredOperations,
               onForbiddenOperation,
@@ -271,8 +276,9 @@ describe('Operation registry plugin', () => {
         const server = new ApolloServerMock({
           typeDefs,
           mockEntireSchema: true,
-          engine: engineOptions,
+          apollo,
           plugins: [
+            ApolloServerPluginUsageReportingDisabled(),
             plugin({
               onForbiddenOperation,
             })(),
